+++ include/pbcopper/data/MappedRead.h
@@ -65,9 +65,10 @@ struct MappedRead : public Read
         GapBehavior gapBehavior = GapBehavior::IGNORE,
         SoftClipBehavior softClipBehavior = SoftClipBehavior::KEEP) const;
 
-    Frames AlignedPulseWidth(Orientation orientation = Orientation::NATIVE,
-                             GapBehavior gapBehavior = GapBehavior::IGNORE,
-                             SoftClipBehavior softClipBehavior = SoftClipBehavior::KEEP) const;
+    boost::optional<Frames> AlignedPulseWidth(
+        Orientation orientation = Orientation::NATIVE,
+        GapBehavior gapBehavior = GapBehavior::IGNORE,
+        SoftClipBehavior softClipBehavior = SoftClipBehavior::KEEP) const;
 
     int32_t NumMismatches() const;
 };